Persist undo limit to disk

+++ JotUI/JotUI/JotViewState.h
@@ -22,7 +22,7 @@
 
 // ability to cancel strokes
 @property(nonatomic, weak) NSObject<JotStrokeDelegate>* delegate;
-@property(nonatomic, assign) int undoLimit;
+@property(nonatomic, assign) NSInteger undoLimit;
 
 // backing textures
 @property(nonatomic, strong) JotGLTexture* backgroundTexture;
